improved imei_automator route

+++ imei_processor.py
@@ -26,18 +26,18 @@ def imei_machine(work_book):
             # Usually phone descriptions end with 1 or 2 so here we filter for phones essentially
             if column_headers is not None and column_headers.endswith("1"):
                 # Assigning the ONLY IMEI to be put through the IMEI info API
-                first_imei = int(cols[1])
+                first_imei = cols[1]
                 print(first_imei)
                 imei_data = imei_finder(first_imei)
                 for cells in clean_cols:
                     # Checking to see if the cell only contains numbers
-                    if isinstance(cells, float):
+                    if isinstance(cells, int):
                         # Appending to the new ws and leaving the second IMEI slot empty
                         new_ws.append([cells, " ", imei_data[1], imei_data[0].upper()])
             # filtering for the second IMEI
             elif column_headers is not None and column_headers.endswith("2"):
                 for cells in cols:
-                    if isinstance(cells, float):
+                    if isinstance(cells, int):
                         # Assigning the second IMEI next to the first one we put into the ws above.
                         new_ws.cell(row=row_number, column=2, value=cells)
                         # Need to keep track of number of rows the s
